Random Walk - Corrected Trajectory

+++ app/src/main/java/com/decagon/pf_montecarlos/ClassModels/RandomWalk.java
@@ -2,14 +2,14 @@ package com.decagon.pf_montecarlos.ClassModels;
 
 public class RandomWalk {
     public int steps;
-    public float[] xArray;
-    public float[] yArray;
+    public int[] xArray;
+    public int[] yArray;
 
     public RandomWalk(int iterations){
         steps = iterations;
 
-        xArray = new float[iterations + 1];
-        yArray = new float[iterations + 1];
+        xArray = new int[iterations + 1];
+        yArray = new int[iterations + 1];
 
         xArray[0] = 0;
         yArray[0] = 0;
